requirements.txt
openai>=1.0.0
elevenlabs>=0.2.0
opencv-python>=4.8.0
pygame>=2.5.0
python-dotenv>=1.0.0
numpy>=1.24.0